Move all tests to tests folder in root
